Moved user checking to checkGameSales
The issue is that I'm using the presence of 'user' to determine
whether this is the daily sales check that's being performed or
if it's a user triggered check. In my last commit, I passed the
user info along to this method from a place that I used to be
passing a null. This caused claptrap to think it was a user triggered
sale check and started posting to slack.

HEAVY REFACTORING REQUIRED